支持String了

+++ branches/Lite-1_0/src/main/org/xidea/el/json/JSONTokenizer.java
@@ -214,7 +214,7 @@ public class JSONTokenizer {
 		String ns = value.substring(begin, start);
 //		System.out.println(ns);
 		if (isFloatingPoint) {
-			return Float.parseFloat(ns);
+			return Double.parseDouble(ns);
 		} else {
 			return Long.parseLong(ns);
 		}
